+++ packages/example/src/components/connector/sdk-connection-provider.tsx
@@ -1,0 +1,43 @@
+import React, { useContext } from "react"
+import { useRxOrThrow } from "@rixio/react"
+import { createRaribleSdk } from "@rarible/sdk"
+import type { ConnectionState } from "@rarible/connector"
+import { IConnector, getStateDisconnected, Connector } from "@rarible/connector"
+import { IRaribleSdk } from "@rarible/sdk/build/domain"
+import { EnvironmentContext } from "./environment-selector-provider"
+import { IWalletAndAddress } from "./wallet-connetion"
+
+export interface IConnectorContext {
+	connector?: IConnector<string, IWalletAndAddress>
+	state: ConnectionState<IWalletAndAddress>
+	sdk?: IRaribleSdk
+	walletAddress?: string
+}
+
+export const ConnectorContext = React.createContext<IConnectorContext>({
+	connector: undefined,
+	state: getStateDisconnected(),
+	sdk: undefined,
+	walletAddress: undefined
+})
+
+export interface ISdkConnectionProviderProps {
+	connector: Connector<string, IWalletAndAddress>
+}
+
+export function SdkConnectionProvider({ connector, children }: React.PropsWithChildren<ISdkConnectionProviderProps>) {
+	const {environment} = useContext(EnvironmentContext)
+	const conn = useRxOrThrow(connector.connection)
+	const sdk = conn.status === "connected" ? createRaribleSdk(conn.connection.wallet, environment) : undefined
+
+	const context: IConnectorContext = {
+		connector,
+		state: conn,
+		sdk,
+		walletAddress: conn.status === "connected" ? conn.connection.address : undefined,
+	}
+
+	return <ConnectorContext.Provider value={context}>
+		{children}
+	</ConnectorContext.Provider>
+}