+++ packages/sdk/src/index.ts
@@ -1,8 +1,7 @@
-import type { BlockchainWallet, WalletByBlockchain } from "@rarible/sdk-wallet"
 import type { ContractAddress } from "@rarible/types"
-import { toContractAddress } from "@rarible/types"
 import type { Maybe } from "@rarible/types/build/maybe"
-import { Blockchain } from "@rarible/api-client/build/models/Blockchain"
+import { Blockchain, BlockchainGroup } from "@rarible/api-client"
+import type { BlockchainWallet, WalletByBlockchain } from "@rarible/sdk-wallet"
 import type { IApisSdk, IRaribleInternalSdk, IRaribleSdk, IRaribleSdkConfig, ISdkContext } from "./domain"
 import { LogsLevel } from "./domain"
 import { getSdkConfig } from "./config"
@@ -29,24 +28,24 @@ export function createRaribleSdk(
 	const apis = createApisSdk(env, config?.apiClientParams)
 	const instance = createUnionSdk(
 		createEthereumSdk(
-			filterWallet(wallet, Blockchain.ETHEREUM),
+			filterWallet(wallet, BlockchainGroup.ETHEREUM),
 			apis,
 			blockchainConfig.ethereumEnv,
 			config?.apiClientParams,
 			config?.logs ?? LogsLevel.TRACE
 		),
 		createFlowSdk(
-			filterWallet(wallet, Blockchain.FLOW),
+			filterWallet(wallet, BlockchainGroup.FLOW),
 			apis,
 			blockchainConfig.flowEnv
 		),
 		createTezosSdk(
-			filterWallet(wallet, Blockchain.TEZOS),
+			filterWallet(wallet, BlockchainGroup.TEZOS),
 			apis,
 			blockchainConfig.tezosNetwork
 		),
 		createEthereumSdk(
-			filterWallet(wallet, Blockchain.POLYGON),
+			filterWallet(wallet, BlockchainGroup.ETHEREUM),
 			apis,
 			blockchainConfig.polygonNetwork,
 			config?.apiClientParams,
@@ -104,7 +103,7 @@ function setupMiddleware(
 	}
 }
 
-function filterWallet<T extends Blockchain>(
+function filterWallet<T extends BlockchainGroup>(
 	wallet: Maybe<BlockchainWallet>,
 	blockchain: T
 ): Maybe<WalletByBlockchain[T]> {
@@ -117,8 +116,7 @@ function filterWallet<T extends Blockchain>(
 function createSell(sell: ISellInternal, apis: IApisSdk): ISell {
 	return async ({ itemId }) => {
 		const item = await apis.item.getItemById({ itemId })
-		const collectionId = toContractAddress(item.contract)
-		const response = await sell({ collectionId })
+		const response = await sell({ blockchain: item.blockchain })
 		return {
 			...response,
 			maxAmount: item.supply,
@@ -135,7 +133,8 @@ function createMintAndSell(mint: IMint, sell: ISellInternal): IMintAndSell {
 	return async request => {
 		const mintResponse = await mint(request)
 		const collectionId = getCollectionId(request)
-		const sellResponse = await sell({ collectionId })
+		const blockchain = getBlockchainCollectionId(collectionId)
+		const sellResponse = await sell({ blockchain })
 
 		const mintAction = mintResponse.submit
 			.around(
@@ -169,6 +168,14 @@ export function getCollectionId(req: HasCollectionId | HasCollection): ContractA
 		return req.collection.id
 	}
 	return req.collectionId
+}
+
+function getBlockchainCollectionId(contract: ContractAddress): Blockchain {
+	const [blockchain] = contract.split(":")
+	if (!(blockchain in Blockchain)) {
+		throw new Error(`Unrecognized blockchain in contract ${contract}`)
+	}
+	return blockchain as Blockchain
 }
 
 type MiddleMintType = {