+++ packages/sdk/src/sdk-blockchains/ethereum/index.ts
@@ -17,10 +17,9 @@ import { EthereumCancel } from "./cancel"
 import { EthereumBalance } from "./balance"
 import { EthereumTokenId } from "./token-id"
 import { EthereumCreateCollection } from "./create-collection"
-import type { EVMBlockchain } from "./common"
 
 export function createEthereumSdk(
-	wallet: Maybe<EthereumWallet<EVMBlockchain>>,
+	wallet: Maybe<EthereumWallet>,
 	apis: IApisSdk,
 	network: EthereumNetwork,
 	params?: ConfigurationParameters,
@@ -28,7 +27,7 @@ export function createEthereumSdk(
 ): IRaribleInternalSdk {
 	const sdk = createRaribleSdk(wallet?.ethereum, network, { apiClientParams: params, logs: logs })
 	const sellService = new EthereumSell(sdk, network)
-	const balanceService = new EthereumBalance(sdk)
+	const balanceService = new EthereumBalance(sdk, network)
 	const bidService = new EthereumBid(sdk, wallet, balanceService, network)
 	const mintService = new EthereumMint(sdk, apis, network)
 	const fillerService = new EthereumFill(sdk, wallet, network)
@@ -56,6 +55,7 @@ export function createEthereumSdk(
 		},
 		balances: {
 			getBalance: balanceService.getBalance,
+			convert: balanceService.convert,
 		},
 		restriction: {
 			canTransfer(): Promise<CanTransferResult> {