+++ packages/sdk/src/sdk-blockchains/flow/sell.ts
@@ -9,6 +9,8 @@ import type * as OrderCommon from "../../types/order/common"
 import type { CurrencyType } from "../../common/domain"
 import { OriginFeeSupport, PayoutsSupport } from "../../types/order/fill/domain"
 import type { IApisSdk } from "../../domain"
+import { getCurrencyAssetType } from "../../common/get-currency-asset-type"
+import type { PrepareSellInternalResponse } from "../../types/order/sell/domain"
 import {
 	convertFlowOrderId,
 	getFlowCollection,
@@ -33,14 +35,14 @@ export class FlowSell {
 		return this.apis.order.getOrderById({ id: request })
 	}
 
-	async sell(request: OrderCommon.PrepareOrderInternalRequest): Promise<OrderCommon.PrepareOrderInternalResponse> {
-		const contract = getFlowCollection(request.collectionId)
+	async sell(): Promise<PrepareSellInternalResponse> {
 		const sellAction = Action.create({
 			id: "send-tx" as const,
 			run: async (sellRequest: OrderCommon.OrderInternalRequest) => {
-				if (sellRequest.currency["@type"] === "FLOW_FT") {
-					const currency = getFungibleTokenName(sellRequest.currency.contract)
-					const { itemId } = parseUnionItemId(sellRequest.itemId)
+				const requestCurrency = getCurrencyAssetType(sellRequest.currency)
+				if (requestCurrency["@type"] === "FLOW_FT") {
+					const currency = getFungibleTokenName(requestCurrency.contract)
+					const { itemId, contract } = parseUnionItemId(sellRequest.itemId)
 					return this.sdk.order.sell({
 						collection: contract,
 						currency,
@@ -50,17 +52,17 @@ export class FlowSell {
 					})
 
 				}
-				throw new Error(`Unsupported currency type: ${sellRequest.currency["@type"]}`)
+				throw new Error(`Unsupported currency type: ${requestCurrency["@type"]}`)
 			},
 		}).after((tx) => convertFlowOrderId(tx.orderId))
 
 
 		return {
-			multiple: false,
 			supportedCurrencies: FlowSell.supportedCurrencies,
 			baseFee: getFlowBaseFee(this.sdk),
 			originFeeSupport: OriginFeeSupport.FULL,
 			payoutsSupport: PayoutsSupport.NONE,
+			supportsExpirationDate: false,
 			submit: sellAction,
 		}
 	}