+++ packages/sdk/CHANGELOG.md
@@ -3,42 +3,101 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [0.8.44-hotfix.3](https://github.com/rarible/sdk/compare/v0.8.44-hotfix.2...v0.8.44-hotfix.3) (2022-04-06)
+## [0.8.51](https://github.com/rarible/sdk/compare/v0.8.49...v0.8.51) (2022-04-01)
+
+
+### Features
+
+* add new addresses for eth ([3b3b5cc](https://github.com/rarible/sdk/commit/3b3b5cc956a0512b55b637b3eebf20dc0d3d3d56))
+* weth contracts was added ([b18b57d](https://github.com/rarible/sdk/commit/b18b57d5c78c85c6227a9fca8ae0337e6988d976))
+
+
+
+
+
+## [0.8.49](https://github.com/rarible/sdk/compare/v0.8.48...v0.8.49) (2022-03-31)
+
+
+### Features
+
+* rm multiple field from sell prepare response ([5381e47](https://github.com/rarible/sdk/commit/5381e47d272e11dbac01a58c7eedae09351ce69a))
+
+
+
+
+
+## [0.8.48](https://github.com/rarible/sdk/compare/v0.8.47...v0.8.48) (2022-03-30)
+
+
+### Features
+
+* currency id support for get balance and get convertable value ([5906797](https://github.com/rarible/sdk/commit/5906797b15d13f71c5c6bf42fd918c7ab10f4d1c))
+
+
+
+
+
+## [0.8.47](https://github.com/rarible/sdk/compare/v0.8.46...v0.8.47) (2022-03-28)
 
 
 ### Bug Fixes
 
-* tezos contracts ([963944d](https://github.com/rarible/sdk/commit/963944d84abed292f5d6ed025767df1facb64688))
+* remove lazy mint on polygon ([#169](https://github.com/rarible/sdk/issues/169)) ([e74bc31](https://github.com/rarible/sdk/commit/e74bc314249af7a208c9d4cbae0b2d1e1c752821))
+* tezos sign function and improve deps ([#158](https://github.com/rarible/sdk/issues/158)) ([791e8c8](https://github.com/rarible/sdk/commit/791e8c8cf2cbac84b77c408993134eae79280049))
 
 
 
 
 
-## [0.8.44-hotfix.2](https://github.com/rarible/sdk/compare/v0.8.44-hotfix.1...v0.8.44-hotfix.2) (2022-04-06)
-
-
-### Bug Fixes
-
-* ethereum test, skip some tezos tests ([046ae1c](https://github.com/rarible/sdk/commit/046ae1c5d398d045d567763d1dc9be132a1f1238))
-
-
-
-
-
-## [0.8.44-hotfix.1](https://github.com/rarible/sdk/compare/v0.8.44-hotfix.0...v0.8.44-hotfix.1) (2022-04-06)
-
-**Note:** Version bump only for package @rarible/sdk
-
-
-
-
-
-## [0.8.44-hotfix.0](https://github.com/rarible/sdk/compare/v0.8.44...v0.8.44-hotfix.0) (2022-04-04)
+## [0.8.46-hotfix.1](https://github.com/rarible/sdk/compare/v0.8.46-hotfix.0...v0.8.46-hotfix.1) (2022-03-21)
 
 
 ### Bug Fixes
 
-* tezos contracts upgrade ([d4ba3aa](https://github.com/rarible/sdk/commit/d4ba3aadb6bd1db8c6267a4b912ff3ede0c23032))
+* rm esm build ([37a20d2](https://github.com/rarible/sdk/commit/37a20d202d15fb18d9d1aff9b7907251f7ad0caf))
+
+
+
+
+
+## [0.8.46-hotfix.0](https://github.com/rarible/sdk/compare/v0.8.46...v0.8.46-hotfix.0) (2022-03-21)
+
+
+### Bug Fixes
+
+* remove lazy mint on polygon ([52a43d3](https://github.com/rarible/sdk/commit/52a43d31a191993daf7ca4084da51ea79a929830))
+
+
+
+
+
+## [0.8.46](https://github.com/rarible/sdk/compare/v0.8.45...v0.8.46) (2022-03-19)
+
+
+### Bug Fixes
+
+* added missing apis to sdk: currency + auctions ([b7c2640](https://github.com/rarible/sdk/commit/b7c264022c838b916e5c6d1976c0e066df47a350))
+* move Example to this monorepo + rework blockchain in EthereumWallet ([#161](https://github.com/rarible/sdk/issues/161)) ([858bacb](https://github.com/rarible/sdk/commit/858bacb094c44fae65cd940ddffc2a3378d8210b))
+
+
+### Features
+
+* convert method ([#149](https://github.com/rarible/sdk/issues/149)) ([d61e2f5](https://github.com/rarible/sdk/commit/d61e2f5fa86754f689ab3656564227cc80d1ac0b))
+* dev env for tezos ([#163](https://github.com/rarible/sdk/issues/163)) ([d5fb00d](https://github.com/rarible/sdk/commit/d5fb00dc654cf331d480d732d9a1066730461cf7))
+* expiration date ([3a0eafd](https://github.com/rarible/sdk/commit/3a0eafd27015cf10f177c27ef5269bf11bf0be2d))
+* polygon example ([552df3f](https://github.com/rarible/sdk/commit/552df3fcb64398ebc1931cac7c7aaa3485e4d421))
+* test mint with token id ([de2a37e](https://github.com/rarible/sdk/commit/de2a37ec066ac7ba8df1c3eea5848d528addd1a3))
+
+
+
+
+
+## [0.8.45](https://github.com/rarible/sdk/compare/v0.8.44...v0.8.45) (2022-03-04)
+
+
+### Bug Fixes
+
+* polygon bid ([#155](https://github.com/rarible/sdk/issues/155)) ([1350930](https://github.com/rarible/sdk/commit/1350930f18b581401e033ad9db0db38eae317945))